add Chimbuko class to run all parts and make simplified main driver
